pulse-server/pyproject.toml
[project]
name = "dietracker-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "psycopg[binary]>=3.2",
    "psycopg-pool>=3.2",
    "sqlalchemy>=2.0",
    "greenlet>=3.0",
    "alembic>=1.16",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "fastmcp>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dietracker_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "integration: marks tests that require a real PostgreSQL database",
]
